debug

+++ src/main/java/run/itlife/service/S3ServiceImpl.java
@@ -30,8 +30,10 @@ public class S3ServiceImpl implements S3Service {
 
     public String uploadS3File(String username, File file) {
         String fileName = "img/users/" + username + "/" + file.getName();
-        //s3Client.putObject(new PutObjectRequest(bucketName, fileName, file));
-        file.delete();
+        if(fileName != null && file != null) {
+            s3Client.putObject(new PutObjectRequest(bucketName, fileName, file));
+            file.delete();
+        }
         return fileName;
     }
